requirements.txt
Django==2.2
django-allauth
django-rest-auth
djangorestframework
django-extensions==2.1.6
psycopg2-binary==2.8.2
pygraphviz==1.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
